# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "modules Platform API",
    "version": "1.0.0",
    "description": "REST API for interacting the modules platform.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-org/modules-repo"
    },
    "license": {
      "name": "Unlicense",
      "url": "https://unlicense.org/"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}/api",
      "description": "Configurable server",
      "variables": {
        "protocol": {
          "enum": ["http", "https"],
          "default": "http"
        },
        "host": {
          "default": "localhost:3000",
          "description": "Server host and port"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session-based authentication using JWT tokens from Better Auth"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for programmatic access. Supports 'read' and 'write' scopes."
      }
    },
    "schemas": {
...
      "ReleaseResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Release ID"
          },
          "moduleId": {
            "type": "string",
            "description": "Module ID"
          },
          "version": {
            "type": "string",
            "description": "Release version"
          },
          "downloadUrl": {
            "type": "string",
            "description": "Download URL"
          },
          "changelog": {
            "type": "string",
            "nullable": true,
            "description": "Release notes"
          },
          "size": {
            "type": "string",
            "nullable": true,
            "description": "File size"
          },
          "downloads": {
            "type": "number",
            "description": "Download count"
          },
          "isLatest": {
            "type": "boolean",
            "description": "Latest release flag"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Release date"
          },
          "assets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "downloadUrl": {
                  "type": "string"
                },
                "size": {
                  "type": "string"
                },
                "contentType": {
                  "type": "string",
                  "nullable": true
                },
                "downloads": {
                  "type": "number"
                }
              },
              "required": ["id", "name", "downloadUrl", "size", "downloads"]
            },
            "nullable": true,
            "description": "Release assets"
          }
        },
        "required": [
          "id",
          "moduleId",
          "version",
          "downloadUrl",
          "downloads",
          "isLatest",
          "createdAt"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/modules/{id}/stats": {
      "get": {
        "operationId": "get-modules-{id}-stats",
        "summary": "Get detailed module statistics",
        "description": "Get statistics for a specific module including downloads, ratings, and trends",
        "tags": ["Modules"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Module ID"
            },
            "required": true,
            "description": "Module ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/modules/{id}/releases": {
      "get": {
        "operationId": "get-modules-{id}-releases",
        "summary": "Get module releases",
        "description": "Retrieve version history and release information for a specific module. Can return all releases or just the latest release with metadata.",
        "tags": ["s Releases"],
        "parameters": [
          {
            "in": "query",
            "name": "latest",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "description": "Return only latest release"
            },
            "required": false,
            "description": "Return only latest release"
          },
          {
            "in": "query",
            "name": "includeAssets",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "description": "Include release assets"
            },
            "required": false,
            "description": "Include release assets"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Module ID"
            },
            "required": true,
            "description": "Module ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved module releases with version history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "releases": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReleaseResponseSchema"
                      },
                      "description": "List of module releases"
                    }
                  },
                  "required": ["releases"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/user/role": {
      "get": {
        "operationId": "get-user-role",
        "summary": "Get user role",
        "description": "Get the role of the authenticated user (user or admin)",
        "tags": ["User"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isAdmin": {
                      "type": "boolean",
                      "description": "Admin role status"
                    },
                    "userId": {
                      "type": "string",
                      "description": "User ID"
                    },
                    "email": {
                      "type": "string",
                      "nullable": true,
                      "description": "User email"
                    }
                  },
                  "required": ["isAdmin", "userId"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  }
}